Performed the Image Editing Processes that Will be Needed to Compare Paintings
Will be using Keras in order to create the neural network for this project. In this instance the image processing that are featured in Keras were used to try to edit a simple monet painting. Will conduct further along for greater understanding.

Need to continue along, and begin feeding in monet paintings as well as other paintings as well.